(#16186) Support listing of values for fact
Previous to this commit, there was no way to use the Inventory Service
to retrieve all the values for a given fact and a list of nodes that
contain each value.

This commit adds an available parameter, **key**, to the Inventory
Service API.  Possible values for the parameter are
* node - The given request is for a node's facts (default)
* fact - The given request is for a fact's values and the nodes that
  have them

By implementing the **key** parameter, the API is fully backwards compatible.
When querying a fact, what's returned is a hash.  The keys are the values of
requested fact. The value for each key is an array containing each node that
has that value.

With this commit, the values for facts such as operatingsystem,
puppetversion, architecture, virtual, etc can be retrieved along with
the popularity of each value by determining the size of the node array.

The following is an example of how to use the key parameter:
`curl -k -H "Accept: yaml"
https://localhost:8140/production/facts/operatingsystem?key=fact`

which will return:

```
---
  Ubuntu:
    - agent1.vagrant.internal
    - agent2.vagrant.internal
    - master.vagrant.internal
  CentOS:
    - test-centos.localdomain
```