Add dependencies feature for package type
Previous to this commit, package providers could not provide dependency
information for installed packages on the node.  This commit introduces
a new feature, dependencies, that providers can implement.  The
'dependencies' feature is a read only property meaning it can not be
used in a manifest.  To implement the feature, a package provider only
has to declare it implements the feature by adding a line to the
provider: `has_feature :dependencies`. In addition, the `dependencies`
method must be implemented and must return an array with each
array element being a string that contains the name of the package that
is a direct dependency of the package resource.

Futher, this commit implements the dependencies feature in the apt
package provider.